Update config.py
hi, Tyler.  Doubt this affects your code, but I need to know the position of each channel in the array for the receptive field mapping function I am writing.  In looking at an mea, I noticed that there are only 59 channels, not 60 (in addition to the 4 empty corners of the array, one other channel is missing along the left-hand side near the middle - channel # 15 using their numbering system [(0,4) in your system].  I changed the channel_map values in my code and deleted item 60.